refactoring takoyakibutton
